adapt next format change by LGA - daily routine ;-)
- new URL for report pickup
- new line spacing inside PDF header, resulting in totally different field order

Try to mitigate this by even more agressive pre-processing, removing more info
I do not need. Adapt regexes once again. With next changes, I'll try to run a
number of regexes on the data, in the hope that there is at least a finite set
of ways to express the header (but humans are always in for a surprise ;-))